Fix DROP Filestorage Database with database name error
DROP Filestorage Database was not working when database name was given.
For example: "DROP FILESTORAGE DATABASE testDBFS" would result in the following error:
"Error: Cannot drop database file, because it does not exist".
This was because FS.dropDatabase (FileStorage) was expecting the filename to be given directly.
However, the method signature for FS.dropDatabase and the documentation for DROP is with
the database name. As a result, when the database name is mentioned along with DROP, it treats
the name as the JSON file.
This commit adds support to DROP Filestorage databases with the database name specified.
The existing behavior of dropping Filestorage databases specifying the filename will still continue to work.